requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.9.0
future==0.18.3
huggingface-hub==0.12.0
idna==3.4
more-itertools==9.0.0
numpy==1.24.1
openai-whisper==20230124
packaging==23.0
pytube==12.1.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
